Allow self referencing content type creation
